+++ libs/langchain/tests/unit_tests/llms/test_openai.py
@@ -1,3 +1,4 @@
+import asyncio
 import os
 from typing import Any
 from unittest.mock import MagicMock, patch
@@ -5,6 +6,10 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from langchain.llms.openai import OpenAI
+from tests.unit_tests.callbacks.fake_callback_handler import (
+    FakeAsyncCallbackHandler,
+    FakeCallbackHandler,
+)
 
 os.environ["OPENAI_API_KEY"] = "foo"
 
@@ -45,44 +50,62 @@ def mock_completion() -> dict:
 
 
 @pytest.mark.requires("openai")
-def test_openai_calls(mock_completion: dict) -> None:
+def test_openai_retries(mock_completion: dict) -> None:
     llm = OpenAI()
     mock_client = MagicMock()
     completed = False
+    raised = False
+    import openai
 
     def raise_once(*args: Any, **kwargs: Any) -> Any:
-        nonlocal completed
+        nonlocal completed, raised
+        if not raised:
+            raised = True
+            raise openai.error.APIError
         completed = True
         return mock_completion
 
     mock_client.create = raise_once
+    callback_handler = FakeCallbackHandler()
     with patch.object(
         llm,
         "client",
         mock_client,
     ):
-        res = llm.predict("bar")
+        res = llm.predict("bar", callbacks=[callback_handler])
         assert res == "Bar Baz"
     assert completed
+    assert raised
+    assert callback_handler.retries == 1
 
 
 @pytest.mark.requires("openai")
+@pytest.mark.asyncio
 async def test_openai_async_retries(mock_completion: dict) -> None:
     llm = OpenAI()
     mock_client = MagicMock()
     completed = False
+    raised = False
+    import openai
 
-    def raise_once(*args: Any, **kwargs: Any) -> Any:
-        nonlocal completed
+    async def araise_once(*args: Any, **kwargs: Any) -> Any:
+        nonlocal completed, raised
+        if not raised:
+            raised = True
+            raise openai.error.APIError
+        await asyncio.sleep(0)
         completed = True
         return mock_completion
 
-    mock_client.create = raise_once
+    mock_client.acreate = araise_once
+    callback_handler = FakeAsyncCallbackHandler()
     with patch.object(
         llm,
         "client",
         mock_client,
     ):
-        res = llm.apredict("bar")
+        res = await llm.apredict("bar", callbacks=[callback_handler])
         assert res == "Bar Baz"
     assert completed
+    assert raised
+    assert callback_handler.retries == 1
